requirements.txt
Django>=4.2
pandas>=2.0
pyarrow>=15.0
duckdb>=0.9
polars<=1.23.0
djangorestframework